test: Refactor new test
